+++ include/pflib/algorithm/toa_vref_scan.h
@@ -15,7 +15,7 @@ namespace pflib::algorithm {
  *
  * @note Only functional for single-ROC targets
  */
-std::map<std::string, std::map<std::string, int>> toa_vref_scan(Target* tgt,
+std::map<std::string, std::map<std::string, uint64_t>> toa_vref_scan(Target* tgt,
                                                                 ROC roc);
 
 }  // namespace pflib::algorithm
